ThreadContext: Replace EmptyIterator with Collections.emptyIterator()
Log4J 2 requires Java 8 now, so using Java 7 classes is possible. It was unclear from the comments and commit messages whether or not the EmptyIterator was used because Collections.emptyIterator() was not available, or because it was deemed unfit at the time.

Switching to the JDK variant results in a slight change in behavior. The EmptyIterator would no-op on a call to remove(). The JDK variant throws an IllegalStateException. This seems acceptable since all other mutator methods of the EmptyThreadContextStack throw an UnsupportedOperationException. While the exceptions aren't the same, it's clear that mutating isn't supported.