LOG4J2-??? Replace internal logging with format strings
Replace trivial internal logging instances that used string concatenation with format strings and arguments. This isn't so much a performance thing as it is setting a good example for downstream projects. Similarly the use of Unbox.box(), especially in test, is not for performance, but so that outside developers reading the code see how to use it, and that it is being used.

It should be noted that there are a handful of string concatenation sites left, that all perform more complex logging, typically involving exceptions, that can be converted separately.